refactor(api):  move findDivisionByOrganizationId controller into src
