+++ api/src/prescription/organization-learner/application/learner-list-controller.js
@@ -1,3 +1,4 @@
+import * as divisionSerializer from '../../campaign/infrastructure/serializers/jsonapi/division-serializer.js';
 import { usecases } from '../domain/usecases/index.js';
 import * as organizationParticipantsSerializer from '../infrastructure/serializers/jsonapi/organization-participants-serializer.js';
 import { mapCertificabilityByLabel } from './../../shared/application/helpers.js';
@@ -22,5 +23,12 @@ const findPaginatedFilteredParticipants = async function (
   return dependencies.organizationParticipantsSerializer.serialize(results);
 };
 
-const learnerListController = { findPaginatedFilteredParticipants };
+const getDivisions = async function (request) {
+  const organizationId = request.params.id;
+  const divisions = await usecases.findDivisionsByOrganization({ organizationId });
+  return divisionSerializer.serialize(divisions);
+};
+
+const learnerListController = { findPaginatedFilteredParticipants, getDivisions };
+
 export { learnerListController };
